LLVM 3.5 build of the Faust package.
Changes the Debian build so that we get a package linked against LLVM 3.5
which can be used as a drop-in replacement for the mainline faust package.
This is needed for better interoperability with the Pure programming language
(https://agraef.github.io/pure-lang/) which doesn't work with newer LLVM
versions yet.